requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.5
djangorestframework==3.14.0
loguru==0.7.2
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
win32-setctime==1.1.0
